order add , signle order, order total price

+++ src/modules/users/user.service.ts
@@ -2,21 +2,33 @@ import { IOrders, IUser } from './user.interface';
 import User from './user.model';
 
 const createUser = async (userData: IUser) => {
+  if (await User.isUserExist(userData.userId)) {
+    throw new Error('User already exists!');
+  }
   const result = await User.create(userData);
   return result;
 };
 
 const getAllUsers = async () => {
-  const result = await User.find();
+  const result = await User.find(
+    {},
+    { username: 1, fullName: 1, age: 1, email: 1, address: 1, _id: 0 },
+  );
   return result;
 };
 
 const getSingleUser = async (userId: number) => {
+  if (!(await User.isUserExist(userId))) {
+    throw new Error('User not found!');
+  }
   const result = await User.findOne({ userId });
   return result;
 };
 
 const updateUser = async (userId: number, updatedData: Partial<IUser>) => {
+  if (!(await User.isUserExist(userId))) {
+    throw new Error('User not found!');
+  }
   const result = await User.findOneAndUpdate({ userId }, updatedData, {
     new: true,
   });
@@ -24,26 +36,64 @@ const updateUser = async (userId: number, updatedData: Partial<IUser>) => {
 };
 
 const deleteUser = async (userId: number): Promise<IUser | null> => {
+  if (!(await User.isUserExist(userId))) {
+    throw new Error('User not found!');
+  }
   const result = await User.findOneAndDelete({ userId });
   return result;
 };
 
-const addOrder = async (userId: number, orderData: IOrders): Promise<void> => {
-  const user = await User.findOne({ userId });
-  if (user) {
-    user.addOrder(orderData);
-    await user.save();
+const addOrder = async (userId: number, orderData: IOrders) => {
+  if (!(await User.isUserExist(userId))) {
+    throw new Error('User not found!');
   }
+  const result = await User.updateOne(
+    { userId },
+    {
+      $push: {
+        orders: { ...orderData },
+      },
+    },
+  );
+  return result;
 };
 
-const getAllOrders = async (userId: number): Promise<IOrders[]> => {
-  const user = await User.findOne({ userId }, { password: 0 });
-  return user ? user.getAllOrders() : [];
+const getAllOrders = async (userId: number) => {
+  if (!(await User.isUserExist(userId))) {
+    throw new Error('User not found!');
+  }
+  const user = await User.findOne({ userId }, { orders: 1, _id: 0 });
+  return user;
 };
 
-const calculateTotalPrice = async (userId: number): Promise<number> => {
-  const user = await User.findOne({ userId });
-  return user ? user.calculateTotalPrice() : 0;
+const calculateTotalPrice = async (userId: number) => {
+  if (!(await User.isUserExist(userId))) {
+    throw new Error('User not found!');
+  }
+  const result = await User.aggregate([
+    {
+      $match: {
+        userId,
+      },
+    },
+    {
+      $unwind: '$orders',
+    },
+    {
+      $group: {
+        _id: null,
+        totalPrice: {
+          $sum: {
+            $multiply: ['$orders.price', '$orders.quantity'],
+          },
+        },
+      },
+    },
+    {
+      $project: { _id: 0, totalPrice: 1 },
+    },
+  ]);
+  return result[0];
 };
 
 export const UserServices = {
